+++ examples/cifar.py
@@ -37,7 +37,7 @@ let the developer focus on the important parts.
 CommandLine:
     python examples/cifar.py --gpu=0 --arch=resnet50
     python examples/cifar.py --gpu=0 --arch=wrn_22 --lr=0.003 --schedule=onecycle --optim=adamw
-    python examples/cifar.py --gpu=1,2 --arch=wrn_22 --lr=0.003 --schedule=onecycle --optim=adamw
+    python examples/cifar.py --gpu=1,2,3 --arch=wrn_22 --lr=0.003 --schedule=onecycle --optim=adamw --batch_size=1800
     python examples/cifar.py --gpu=1,2 --arch=resnet50 --lr=0.003 --schedule=onecycle --optim=adamw
 
 """
