+++ netharn/__init__.py
@@ -3,7 +3,7 @@
 # (Remote the --dry to actually run it)
 mkinit netharn --noattrs --dry
 """
-__version__ = '0.1.4.dev0'
+__version__ = '0.1.5.dev0'
 
 from netharn.device import (XPU,)
 from netharn.fit_harn import (FitHarn,)
