--- netharn/util/nms/__init__.py
@@ -1,8 +1,0 @@
-"""
-dont autogen this file
-"""
-# flake8: noqa
-__all__ = ['non_max_supression']
-
-from netharn.util.nms import nms_core
-from netharn.util.nms.nms_core import (non_max_supression,)
