+++ run_developer_setup.sh
@@ -5,6 +5,10 @@ pip install -r requirements.txt
 # Install netharn in developer mode
 pip install -e .
 
+python setup.py clean
+GCC_VERSION=6
+CC=gcc-$GCC_VERSION CXX=g++-$GCC_VERSION CUDAHOSTCXX=g++-$GCC_VERSION python setup.py develop
+
 # Compile C extensions to improve runtime
 python setup.py build_ext --inplace
